services/core/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
orjson>=3.9.0
requests>=2.31.0
redis>=5.0.0
